+++ packages/core/src/browser/window/default-window-service.ts
@@ -19,7 +19,8 @@ import { Event, Emitter } from '../../common';
 import { CorePreferences } from '../core-preferences';
 import { ContributionProvider } from '../../common/contribution-provider';
 import { FrontendApplicationContribution, FrontendApplication } from '../frontend-application';
-import { DEFAULT_WINDOW_HASH, WindowService } from './window-service';
+import { WindowService } from './window-service';
+import { DEFAULT_WINDOW_HASH } from '../../common/window';
 
 @injectable()
 export class DefaultWindowService implements WindowService, FrontendApplicationContribution {