+++ packages/core/src/electron-main/electron-main-application.ts
@@ -30,10 +30,9 @@ import { ContributionProvider } from '../common/contribution-provider';
 import { ElectronSecurityTokenService } from './electron-security-token-service';
 import { ElectronSecurityToken } from '../electron-common/electron-token';
 import Storage = require('electron-store');
-// eslint-disable-next-line @theia/runtime-import-check
-import { DEFAULT_WINDOW_HASH } from '../browser/window/window-service';
 import { isOSX, isWindows } from '../common';
 import { RequestTitleBarStyle, Restart, TitleBarStyleAtStartup, TitleBarStyleChanged } from '../electron-common/messaging/electron-messages';
+import { DEFAULT_WINDOW_HASH } from '../common/window';
 
 const createYargs: (argv?: string[], cwd?: string) => Argv = require('yargs/yargs');